oze-agent/requirements.txt
# Telegram
python-telegram-bot>=21.0,<22.0

# API
fastapi>=0.115.0,<1.0
uvicorn>=0.30.0,<1.0

# AI
anthropic>=0.40.0,<1.0
openai>=1.50.0,<2.0

# Database
supabase>=2.0.0,<3.0

# Google APIs
google-api-python-client>=2.100.0,<3.0
google-auth-oauthlib>=1.2.0,<2.0
google-auth-httplib2>=0.2.0,<1.0

# Security
cryptography>=42.0.0,<45.0
PyJWT>=2.8.0,<3.0

# Scheduler
APScheduler>=3.10.0,<4.0

# Utilities
python-dotenv>=1.0.0,<2.0
httpx>=0.27.0,<1.0
python-multipart>=0.0.9,<1.0
reportlab>=4.2.0,<5.0

# Monitoring
sentry-sdk[fastapi]>=2.0.0,<3.0

# Testing
pytest>=8.0.0,<9.0
pytest-asyncio>=0.23.0,<1.0
pytest-httpx>=0.30.0,<1.0
